Add rate limiting to express server
A rate limiting middleware, express-rate-limit, was added to the express server which helps in limiting the rate of client-side requests. This it does by limiting each IP to 100 requests every 15 minutes, to prevent server overload and increase security. The package 'express-rate-limit' was also added to the dependencies in the package.json file.